[REFACTOR] : BottomSheetContent 파일 분리
